Move palette components to own file for reuse
