+++ app/model/editions/EditionsCard.scala
@@ -43,10 +43,10 @@ case class CardMetadata(
   overrideArticleMainMedia: Option[Boolean],
   coverCardImages: Option[CoverCardImages],
   promotionMetric: Option[Double],
-  bio: Option[String] = None,
-  backgroundHex: Option[String] = None,
-  foregroundHex: Option[String] = None,
-  paletteId: Option[String] = None,
+  bio: Option[String] = None, // Chef
+  backgroundHex: Option[String] = None, // Chef
+  foregroundHex: Option[String] = None, // Chef
+  paletteId: Option[String] = None, // Chef
 )
 
 object CardMetadata {
