+++ fronts-client/src/components/form/ChefMetaForm.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useCallback } from 'react';
 import {
   reduxForm,
   Field,
@@ -33,13 +33,13 @@ import ButtonDefault from 'components/inputs/ButtonDefault';
 import { useDispatch } from 'react-redux';
 import { startOptionsModal } from 'actions/OptionsModal';
 import {
-  ChefPalette,
   ChefPaletteId,
   chefPalettes,
 } from 'constants/feastPalettes';
 import { styled } from 'constants/theme';
 import { noop } from 'lodash';
 import { OptionsModalBodyProps } from 'types/Modals';
+import { PaletteItem } from './Palette';
 
 interface FormProps {
   card: Card;
@@ -113,9 +113,9 @@ const Form = ({
           <ImageRowContainer size={size}>
             <Row>
               <ImageCol>
-                <InputLabel htmlFor="image">Trail image</InputLabel>
+                <InputLabel htmlFor="replaceImage">Trail image</InputLabel>
                 <Field
-                  name={'image'}
+                  name="replaceImage"
                   component={InputImage}
                   criteria={defaultCardTrailImageCriteria}
                 />
@@ -167,8 +167,8 @@ export const ChefMetaForm = ({ cardId, form, ...rest }: ChefMetaFormProps) => {
   const chef = useSelector((state: State) =>
     selectors.selectChefFromCard(state, cardId)
   );
-  const paletteId = useSelector((state: State) =>
-    valueSelector(state, 'paletteId')
+  const paletteId = useSelector(
+    (state: State) => valueSelector(state, 'paletteId') as ChefPaletteId
   );
 
   const dispatch = useDispatch();
@@ -191,7 +191,7 @@ export const ChefMetaForm = ({ cardId, form, ...rest }: ChefMetaFormProps) => {
       chef={chef}
       chefWithoutOverrides={chefWithoutOverrides}
       card={card}
-      initialValues={chef}
+      initialValues={card.meta}
       openPaletteModal={openPaletteModal}
       currentPaletteId={paletteId}
       form={form}
@@ -199,6 +199,11 @@ export const ChefMetaForm = ({ cardId, form, ...rest }: ChefMetaFormProps) => {
     />
   );
 };
+
+const PaletteList = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+`;
 
 const createPaletteForm =
   (formName: string) =>
@@ -247,69 +252,3 @@ const createPaletteForm =
       </PaletteList>
     );
   };
-
-const PaletteItem = ({
-  id,
-  palette,
-  onClick = noop,
-  isSelected = false,
-}: {
-  id: ChefPaletteId;
-  palette: ChefPalette;
-  onClick?: (paletteId: ChefPaletteId) => void;
-  isSelected?: boolean;
-}) => {
-  return (
-    <PaletteOption key={id} onClick={() => onClick(id)} isSelected={isSelected}>
-      <PaletteHeading>{id}</PaletteHeading>
-      <PaletteSwatch colors={[palette.backgroundHex, palette.foregroundHex]} />
-    </PaletteOption>
-  );
-};
-
-const PaletteHeading = styled.h3`
-  margin-top: 0;
-`;
-
-const PaletteOption = styled.div<{ isSelected: boolean }>`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-width: 100px;
-  width: min-content;
-  padding: 10px;
-  margin: 5px 5px 5px 0;
-  border: 2px solid ${({ isSelected }) => (isSelected ? 'darkblue' : '#ccc')};
-  border-radius: 5px;
-  cursor: pointer;
-`;
-
-const PaletteList = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-`;
-
-const PaletteSwatch = ({ colors }: { colors: string[] }) => (
-  <PaletteContainer borderColor={colors[0]}>
-    {colors.map((color) => (
-      <PaletteColor key={color} color={color} />
-    ))}
-  </PaletteContainer>
-);
-
-const PaletteContainer = styled.div<{ borderColor: string | undefined }>`
-  display: flex;
-  width: 50px;
-  height: 50px;
-  border-radius: 5px;
-  border: ${({ borderColor }) =>
-    borderColor ? `2px solid ${borderColor}` : 'none'};
-  overflow: hidden;
-`;
-
-const PaletteColor = styled.div<{ color: string }>`
-  flex-grow: 1;
-  background-color: ${({ color }) => color};
-  height: 100%;
-  min-width: 10px;
-`;
